Clean up relationship types
Removes use of respond_to? in favor of explicit relationship types. i.e. if we say `attributes :foo_ids`, we explicitly intend to only call `object.foo_ids`, not `object.foo.map(&:id)` or equivalent. if you don't specify a relationship in this way, it is probably a good idea to preload associations to avoid N+1s